create new luminosity-focused file/tweak documentation
